Bug fixing in smoothStop
Function was setting a variable acceleration for slow down that could be very slow, depending on walking ratio (3.1 ratio generates one third of 700 mm per sec squared). Further, it mis-computed the expected stop time (which assumed a 700 mm per sec sq. acceleration) so we might not wait long enough until stop time.
This may be the source of stopping errors.